开发计划

+++ cyj-common/src/main/java/com/example/cyjcommon/entity/bean/AppService.java
@@ -40,7 +40,7 @@ public class AppService extends Model<AppService> {
     private int status;
 
     @TableField("sort_code")
-    private int sortCode;
+    private long sortCode;
 
     @TableField("created_date")
     private LocalDateTime createdDate;
