+++ cyj-common/src/main/java/com/example/cyjcommon/entity/bean/Enterprise.java
@@ -31,7 +31,7 @@ public class Enterprise extends Model<Enterprise> {
     private int status;
 
     @TableField("sort_code")
-    private int sortCode;
+    private long sortCode;
 
     @TableField("created_date")
     private LocalDateTime createdDate;
