+++ cyj-common/src/main/java/com/example/cyjcommon/entity/bean/Persistent.java
@@ -37,7 +37,7 @@ public class Persistent extends Model<Persistent> {
     private int status;
 
     @TableField("sort_code")
-    private int sortCode;
+    private long sortCode;
 
     @TableField("created_date")
     private LocalDateTime createdDate;
